refactor(Coin): move out from PVZ.h
