+++ pvzclass/pvzclass.vcxproj.filters
@@ -860,5 +860,8 @@
     <ClInclude Include="Classes\Griditem.hpp">
       <Filter>头文件\Classes</Filter>
     </ClInclude>
+    <ClInclude Include="Classes\GameObject.hpp">
+      <Filter>头文件\Classes</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
